+++ Release Notes.md
@@ -6,6 +6,10 @@
 
 * Fixed app crashing when tapped "Recommend to a friend" on iPad.
 
+### Code improvements
+
+* Updated Fabrics.
+
 ## 1.0.1 (15)
 
 ### New Features